refactor(api-java,api): V10 — W2 nutrition normalize (학습 자료)
audit (P4 — 미사용 W2 placeholder 정합성):
- meal_food_kind ENUM 의 한글값은 i18n / 코드 검색에 불리.
- pet_food.{protein,carbs,fat} 단위 표기 모호 (1회분? per 100g?) — kcal_per_100g
  와 비교 시 일관성 결여.
- 영양정보 컬럼이 DOUBLE PRECISION → 부동소수점 누적 오차.

V10 migration:
- ALTER TYPE meal_food_kind RENAME VALUE  (PG 10+)
    '사료'  → 'kibble'
    '간식'  → 'treat'
    '일반식' → 'regular'
    '처방식' → 'prescription'
  · in-place rename — DROP/CREATE 불요, 데이터 보존.
- ALTER TABLE pet_food RENAME COLUMN
    protein → protein_per_100g
    carbs   → carbs_per_100g
    fat     → fat_per_100g
- ALTER TABLE … ALTER COLUMN … TYPE NUMERIC(10, 2)
    meal:     qty_g, kcal, protein_g, carbs_g, fat_g
    pet_food: kcal_per_100g, protein_per_100g, carbs_per_100g, fat_per_100g
  · PG implicit cast → 데이터 손실 없음.

backend:
- MealFoodKind.java enum 식별자 (사료/간식/일반식/처방식) → 영문 키.
- Meal/PetFood 엔티티 Double → BigDecimal (NUMERIC ↔ BigDecimal 매핑).
- PetFood 컬럼명 + 필드명 통일 (protein_per_100g 등).
- Python: meal.py / pet_food.py 의 sa.Float → sa.Numeric, Mapped[float] →
  Mapped[Decimal]. MealFoodKind enum 값 영문화.

contract / 데이터:
- meal/pet_food 는 service/route 미참조 (placeholder) → 데이터 영향 0.
- DTO / OpenAPI 변경 없음 (외부 노출 0).

학습 포인트
- ALTER TYPE … RENAME VALUE 로 ENUM 값 in-place 변경 가능 (PG 10+).
- DOUBLE PRECISION 의 부동소수점 누적 오차 vs NUMERIC 의 정밀도.
- BigDecimal / Decimal 은 측정/통화 도메인의 표준.

verification:
- ./gradlew test : 81/0 green.
- ./gradlew exportOpenApi : 변동 없음 (DTO 동일).
- pytest apps/api/tests : 95/95 green.

+++ apps/api-java/src/main/java/com/petfinect/api/domain/Meal.java
@@ -1,5 +1,6 @@
 package com.petfinect.api.domain;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
@@ -58,19 +59,20 @@ public class Meal {
 	@Column(name = "food_name", nullable = false)
 	private String foodName;
 
+	// V10: DOUBLE PRECISION → NUMERIC(10, 2) — 영양정보 정밀도 (BigDecimal).
 	@Column(name = "qty_g", nullable = false)
-	private Double qtyG;
+	private BigDecimal qtyG;
 
-	private Double kcal;
+	private BigDecimal kcal;
 
 	@Column(name = "protein_g")
-	private Double proteinG;
+	private BigDecimal proteinG;
 
 	@Column(name = "carbs_g")
-	private Double carbsG;
+	private BigDecimal carbsG;
 
 	@Column(name = "fat_g")
-	private Double fatG;
+	private BigDecimal fatG;
 
 	private String note;
